Rollo.py interfaced to NumericStringParser.py, bot.py used for testing
NumericStringParser reads a math expression as string and return the result of it. Rollo get a string representing a math expression with dices to roll. Rollo converts this keywords ( ex converts 2D20 in two random integers from 1 to 20) and sends a math expression with numbers to NumeriStringParser. To Do: add "1"on the left of D<faces> only keywords; manage the keywords "vantaggio , svantaggio"  to perform double roll; program Discord bot.